Add support for discounts in transaction reporting
A new nullable `Discount` property of type `DiscountDto?` was
added to the `CategoryItemDtos` class to represent discount
information.

In `TransactionsReportingService.cs`, the `Discount` property
is now populated during data mapping. If a `DiscountId` exists
and the `Discount` object is not null, a `DiscountDto` is
created and initialized with relevant properties. If no
discount is applicable, the `Discount` property is set to
`null`.

Additionally, the `Items` property is explicitly omitted in
the mapping process, with a comment indicating that null
values may be ignored by the serializer.